rename tag-groups folder to tags
