+++ src/front/ui/components/tags/index.js
