+++ src/front/ui/components/tags/style.scss
